Fix bookmarks test by pinning window
The tap syncs from the start_date/bookmark until now() and we don't have
new test data on the test repo. This causes tests to fail after a while.
The ability to shift the bookmark, on a per-stream basis, was already put
in place so we use it to make the state work with the data we have.

`Commits` only has two days of data, so we need to put the start date
before the first day and the bookmark on the second day in order for the
second sync to pull less data

---------

Co-authored-by: Dylan Sprayberry <28106103+dsprayberry@users.noreply.github.com>